refactor(113179): refatora varios campos
